// src/app/business-logic/model/tasks/tasksDeleteManyResponse.ts
/**
 * tasks
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 2.13
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { TasksDeleteManyResponseSucceeded } from '././tasksDeleteManyResponseSucceeded';
import { TasksResetManyResponseFailed } from '././tasksResetManyResponseFailed';


export interface TasksDeleteManyResponse {
    succeeded?: Array<TasksDeleteManyResponseSucceeded>;
    failed?: Array<TasksResetManyResponseFailed>;
}